boxcars

Enum Attribute

pub enum Attribute {

    Boolean(bool),
    Byte(u8),
    AppliedDamage(AppliedDamage),
    DamageState(DamageState),
    CamSettings(Box<CamSettings>),
    ClubColors(ClubColors),
    Demolish(Box<Demolish>),
    DemolishExtended(Box<DemolishExtended>),
    DemolishFx(Box<DemolishFx>),
    Enum(u16),
    Explosion(Explosion),
    ExtendedExplosion(ExtendedExplosion),
    FlaggedByte(bool, u8),
    ActiveActor(ActiveActor),
    Float(f32),
    GameMode(u8, u8),
    Int(i32),
    Int64(i64),
    Loadout(Box<Loadout>),
    TeamLoadout(Box<TeamLoadout>),
    Location(Vector3f),
    MusicStinger(MusicStinger),
    PlayerHistoryKey(u16),
    Pickup(Pickup),
    PickupNew(PickupNew),
    QWord(u64),
    Welded(Welded),
    Title(bool, bool, u32, u32, u32, u32, u32, bool),
    TeamPaint(TeamPaint),
    RigidBody(RigidBody),
    String(String),
    UniqueId(Box<UniqueId>),
    Reservation(Box<Reservation>),
    PartyLeader(Option<Box<UniqueId>>),
    PrivateMatch(Box<PrivateMatchSettings>),
    LoadoutOnline(Vec<Vec<Product>>),
    LoadoutsOnline(LoadoutsOnline),
    StatEvent(StatEvent),
    Rotation(Rotation),
    RepStatTitle(RepStatTitle),
    PickupInfo(PickupInfo),
    Impulse(Impulse),
    ReplicatedBoost(ReplicatedBoost),
    LogoData(LogoData),

}

The attributes for updated actors in the network data.

The vast majority of attributes in the network data are rigid bodies. As a performance improvent, any attribute variant larger than the size of a rigid body is moved to the heap (ie: Box::new). This change increased throughput by 40%.

Variants

Boolean(bool)

Byte(u8)

AppliedDamage(AppliedDamage)

DamageState(DamageState)

CamSettings(Box<CamSettings>)

ClubColors(ClubColors)

Demolish(Box<Demolish>)

DemolishExtended(Box<DemolishExtended>)

DemolishFx(Box<DemolishFx>)

Enum(u16)

Explosion(Explosion)

ExtendedExplosion(ExtendedExplosion)

FlaggedByte(bool, u8)

ActiveActor(ActiveActor)

Float(f32)

GameMode(u8, u8)

Int(i32)

Int64(i64)

Loadout(Box<Loadout>)

TeamLoadout(Box<TeamLoadout>)

Location(Vector3f)

MusicStinger(MusicStinger)

PlayerHistoryKey(u16)

Pickup(Pickup)

PickupNew(PickupNew)

QWord(u64)

Welded(Welded)

Title(bool, bool, u32, u32, u32, u32, u32, bool)

TeamPaint(TeamPaint)

RigidBody(RigidBody)

String(String)

UniqueId(Box<UniqueId>)

Reservation(Box<Reservation>)

PartyLeader(Option<Box<UniqueId>>)

PrivateMatch(Box<PrivateMatchSettings>)

LoadoutOnline(Vec<Vec<Product>>)

LoadoutsOnline(LoadoutsOnline)

StatEvent(StatEvent)

Rotation(Rotation)

RepStatTitle(RepStatTitle)

PickupInfo(PickupInfo)

Impulse(Impulse)

ReplicatedBoost(ReplicatedBoost)

LogoData(LogoData)